+++ src/routes/unit.ts
@@ -2,7 +2,6 @@ import { Hono } from "hono";
 import * as unitController from "$controllers/rest/UnitController";
 import * as AuthMiddleware from "$middlewares/authMiddleware";
 import * as unitValidation from "$validations/UnitValidation";
-import { Roles } from "@prisma/client";
 
 const unitRoutes = new Hono();
 
@@ -13,14 +12,15 @@ unitRoutes.get("/:id", unitController.getById);
 unitRoutes.post(
   "/",
   AuthMiddleware.checkJwt,
-  AuthMiddleware.checkRole([Roles.ADMIN]),
+  // AuthMiddleware.checkRole([Roles.ADMIN]),
+  AuthMiddleware.checkAccess("UNIT", "create"),
   unitValidation.validateUnitCreateDTO,
   unitController.create
 );
 unitRoutes.post(
   "/petugas",
   AuthMiddleware.checkJwt,
-  AuthMiddleware.checkRole([Roles.KEPALA_PETUGAS_UNIT]),
+  // AuthMiddleware.checkRole([Roles.KEPALA_PETUGAS_UNIT]),
   unitValidation.validateAddPetugasDTO,
   unitController.addPetugas
 );
@@ -28,7 +28,8 @@ unitRoutes.post(
 unitRoutes.delete(
   "/:nama_unit/petugas",
   AuthMiddleware.checkJwt,
-  AuthMiddleware.checkRole([Roles.KEPALA_PETUGAS_UNIT]),
+  // AuthMiddleware.checkRole([Roles.KEPALA_PETUGAS_UNIT]),
+
   unitValidation.validateRemovePetugasDTO,
   unitController.removePetugas
 );
@@ -36,6 +37,8 @@ unitRoutes.delete(
 unitRoutes.put(
   "/:id",
   unitValidation.validateUnitUpdateDTO,
+
+  AuthMiddleware.checkAccess("UNIT", "update"),
   AuthMiddleware.checkJwt,
   unitController.update
 );
@@ -43,6 +46,7 @@ unitRoutes.put(
 unitRoutes.delete(
   "/",
   unitValidation.validationDeletedUnit,
+  AuthMiddleware.checkAccess("UNIT", "delete"),
   AuthMiddleware.checkJwt,
   unitController.deleteByIds
 );
